# apps/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kedai Bunda API",
        "description": "API untuk Sistem Kasir dan Manajemen Stok Kedai Bunda",
        "contact": {
            "email": "admin@kedaibunda.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/konversi-bahan": {
            "get": {
                "tags": [
                    "Konversi Bahan"
                ],
                "summary": "Ambil semua konversi bahan",
                "operationId": "ba42c071c20333c594b3e6c4ab0892d1",
                "parameters": [
                    {
                        "name": "bahan_baku_id",
                        "in": "query",
                        "description": "Filter by bahan baku ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data konversi bahan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "bahan_baku_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "satuan_asal": {
                                                        "type": "string",
                                                        "example": "ekor"
                                                    },
                                                    "satuan_tujuan": {
                                                        "type": "string",
                                                        "example": "potong"
                                                    },
                                                    "nilai_konversi": {
                                                        "type": "number",
                                                        "example": 8
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Konversi Bahan"
                ],
                "summary": "Tambah konversi bahan baru",
                "operationId": "ff9718fc6e880080b9af32756405d6ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bahan_baku_id",
                                    "satuan_konversi",
                                    "nilai_konversi"
                                ],
                                "properties": {
                                    "bahan_baku_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "satuan_konversi": {
                                        "type": "string",
                                        "example": "potong"
                                    },
                                    "nilai_konversi": {
                                        "type": "number",
                                        "example": 8
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "1 ekor = 8 potong"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Konversi berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/konversi-bahan/{id}": {
            "get": {
                "tags": [
                    "Konversi Bahan"
                ],
                "summary": "Menampilkan detail konversi bahan",
                "operationId": "385086edb958fd4277f1d0b2e26d538b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Konversi Bahan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail konversi bahan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Berhasil mengambil detail konversi bahan"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "bahan_baku_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "satuan_konversi": {
                                                    "type": "string",
                                                    "example": "potong"
                                                },
                                                "nilai_konversi": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 8
                                                },
                                                "keterangan": {
                                                    "type": "string",
                                                    "example": "1 ekor = 8 potong"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konversi bahan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Konversi bahan tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Konversi Bahan"
                ],
                "summary": "Mengupdate konversi bahan",
                "operationId": "8e1468f04d9501688ed81be9802d318f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Konversi Bahan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data konversi bahan yang ingin diupdate (semua field opsional)",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "satuan_konversi": {
                                        "type": "string",
                                        "example": "potong"
                                    },
                                    "nilai_konversi": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "1 ekor = 10 potong kecil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konversi bahan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Konversi bahan berhasil diupdate"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konversi bahan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Konversi bahan tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Konversi Bahan"
                ],
                "summary": "Menghapus konversi bahan",
                "operationId": "be24ceb27f7379bbd1b513466f78bd7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Konversi Bahan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konversi bahan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Konversi bahan berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konversi bahan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Konversi bahan tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}